+++ src/utils.rs
@@ -1,6 +1,6 @@
 
 
 
-pub fn float_compare(a: f32, b: f32) -> bool{
+pub fn float_compare(a: f64, b: f64) -> bool{
     (a - b).abs() < 0.0001
 }
